Fixes for ISPConfig 3.3 Patch 2

Closes #6869, #6884, #6885, #6889, #6890, and #6881

See merge request ispconfig/ispconfig3!2025

+++ interface/web/admin/users_del.php
@@ -47,6 +47,8 @@ $app->auth->check_module_permissions('admin');
 $app->auth->check_security_permissions('admin_allow_del_cpuser');
 if($conf['demo_mode'] == true && $_REQUEST['id'] <= 3) $app->error('This function is disabled in demo mode.');
 
+if(!$app->auth->is_admin()) die('Allowed for administrators only.');
+  
 $app->uses("tform_actions");
 $app->tform_actions->onDelete();
